Move `create_log` into its own util file.
